pyproject.toml
[project]
name = "whatsapp-automation"
version = "0.1.0"
description = "FastAPI webhook receiver for WhatsApp Cloud API with Postgres persistence and local media storage."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16.5",
  "claude-agent-sdk>=0.1.52",
  "fastapi>=0.116.1",
  "httpx>=0.28.1",
  "psycopg[binary]>=3.2.10",
  "pydantic-settings>=2.10.1",
  "sqlalchemy>=2.0.43",
  "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
  "pytest>=8.4.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/whatsapp_automation"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
